+++ packages/@tinacms/toolkit/src/packages/fields/components/Select.tsx
@@ -40,7 +40,7 @@ export interface SelectProps {
 }
 
 export const selectFieldClasses =
-  'shadow appearance-none bg-white text-gray-600 block pl-3 pr-7 py-2 truncate w-full text-base cursor-pointer border border-gray-100 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm rounded-md'
+  'shadow appearance-none bg-white block pl-3 pr-7 py-2 truncate w-full text-base cursor-pointer border border-gray-200 focus:outline-none focus:ring-blue-500 focus:border-blue-500 sm:text-sm rounded-md'
 
 export const Select: React.FC<SelectProps> = ({ input, field, options }) => {
   const selectOptions = options || field.options
@@ -50,7 +50,9 @@ export const Select: React.FC<SelectProps> = ({ input, field, options }) => {
         id={input.name}
         value={input.value}
         onChange={input.onChange}
-        className={selectFieldClasses}
+        className={`${selectFieldClasses} ${
+          input.value ? 'text-gray-700' : 'text-gray-300'
+        } }`}
         {...input}
       >
         {selectOptions ? (