+++ packages/@tinacms/toolkit/src/packages/fields/plugins/ButtonToggleFieldPlugin.tsx
@@ -16,22 +16,12 @@ limitations under the License.
 
 */
 
-export type AnyField = Field & { [key: string]: any }
+import { wrapFieldsWithMeta } from './wrapFieldWithMeta'
+import { ButtonToggle } from '../components'
 
-export interface Field<F extends Field = AnyField> {
-  name: string
-  label?: string | boolean
-  description?: string
-  component: React.FC<any> | string | null
-  inlineComponent?: React.FC<any>
-  parse?: (value: any, name: string, field: F) => any
-  format?: (value: any, name: string, field: F) => any
-  validate?(
-    value: any,
-    allValues: any,
-    meta: any,
-    field: Field
-  ): string | object | undefined
-  defaultValue?: any
-  fields?: F[]
+export const ButtonToggleField = wrapFieldsWithMeta(ButtonToggle)
+
+export const ButtonToggleFieldPlugin = {
+  name: 'button-toggle',
+  Component: ButtonToggleField,
 }