+++ packages/tinacms/src/admin/components/Page.tsx
@@ -12,7 +12,7 @@ limitations under the License.
 */
 
 import React from 'react'
-import { LocalWarning } from '@tinacms/toolkit'
+import { LocalWarning, BillingWarning } from '@tinacms/toolkit'
 
 export const PageWrapper = ({
   children,
@@ -35,6 +35,7 @@ export const PageHeader = ({
 }) => (
   <>
     {isLocalMode && <LocalWarning />}
+    {!isLocalMode && <BillingWarning />}
     <div className="bg-white pb-4 pt-16 border-b border-gray-200 px-12">
       <div className="w-full mx-auto max-w-screen-xl">
         <div className="w-full flex justify-between items-end">{children}</div>