+++ packages/@tinacms/graphql/CHANGELOG.md
@@ -1,5 +1,14 @@
 # tina-graphql
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [7554ea362]
+- Updated dependencies [4ebc44068]
+  - @tinacms/schema-tools@1.1.0
+  - @tinacms/mdx@1.0.4
+
 ## 1.0.3
 
 ### Patch Changes