+++ packages/@tinacms/mdx/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tinacms/mdx
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [7554ea362]
+- Updated dependencies [4ebc44068]
+  - @tinacms/schema-tools@1.1.0
+
 ## 1.0.3
 
 ### Patch Changes